Extracted DrawableHorizontalLine to it's own file.
